Implement windows lockfile
